Changes the way connections are added and become active
Connections now are added to an array that holds the info needed to uniquely identify them. This is done using an EventType called playerJoinInactive, since the player is not yet in the group of active participants in the lobby. At this point the player can only see who is active. Once the player joins, another event is sent with playerJoinActive, and the isActive flag is set to true. After a player connects the list of active players is sent back to that player, and a similar interaction takes place when a player becomes active.